requirements.txt
torch==1.12.1
torchvision==0.13.1
uvicorn==0.18.3
requests==2.28.1
fastapi==0.85.1
pytest==7.2
starlette==0.20.4
python-multipart==0.0.5
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.6.0
Pillow~=9.2.0
pydantic~=1.10.2
pytest==7.2.0
pytest-xdist==3.0.2